feat(day5): implemented part 2 correctly when no overlap exists

+++ day5/solution.h
@@ -12,7 +12,7 @@ namespace solution {
 // Day 5 Solution
 // ============================================================================
 
-using Result = int;
+using Result = long;
 using Ingredient = long;
 using Ingredients = std::vector<Ingredient>;
 template <typename T> struct Range {
